feat: implement calculator function for basic arithmetic operations
Add implementation for the following function:
add(a,b): returns the sum of two numbers.
subtract(a,b): returns the different between two numbers.
sum(array): returns the sum of all the element with the array using 'reduce'.
multiply(array): returns the product of all the element with the array using 'reduce'.
power(a,b): returns 'a' raised to the power of 'b' using 'Math.pow'.
factorial(a,b): returns the factorial of non-negative integer using 'loop'.

These implementation are design to pass all related jest unit test as defined in the spec file.